server/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
twilio